feat: enforce hot-only recency risk controls
Hot-only signals need explicit fade protection so a short-term edge does not expand drawdowns when conditions shift. This change makes recency handling deterministic using executed trade outcomes and persisted pause state.

Updated assess-multi to evaluate consecutive-loss and rolling-loss pause triggers, persist hot-only state, and require reactivation criteria before allowing hot-only recommendations again. Added support for fractional queued qty so reduced-size hot-only approvals can be executed directly.

Assessment reports now include executed-trade fade metrics and pause status for auditability. Unrelated dependency and output changes were intentionally left out of this commit.

+++ journal.py
@@ -445,7 +445,7 @@ def enqueue_signal(
     idea: dict,
     order_type: str,
     limit_price: float | None,
-    qty: int,
+    qty: float,
 ) -> str:
     signal_id = str(uuid4())
     row = {
